+++ packages/core/src/index.ts
@@ -9,6 +9,8 @@ export { parseJson, JsonParseError } from "./json/parse";
 export type { ParsedJson } from "./json/parse";
 export { flatten } from "./flatten";
 export type { FlattenOptions } from "./flatten";
+export { findFiles, globToRegExp } from "./glob";
+export type { FindFilesOptions } from "./glob";
 export { compareKeys } from "./key-comparator/compare";
 export type { KeyDiff } from "./key-comparator/compare";
 export {